split away gateway logic from app.js
